backend/src/api/middleware/auth.py
"""
Authentication middleware (placeholder for future implementation)
"""

from typing import Callable

from fastapi import Request, Response
from starlette.middleware.base import BaseHTTPMiddleware
import structlog

logger = structlog.get_logger("middleware.auth")


class AuthenticationMiddleware(BaseHTTPMiddleware):
    """
    Authentication middleware (placeholder implementation)
    """
    
    async def dispatch(self, request: Request, call_next: Callable) -> Response:
        """
        Process authentication (placeholder)
        
        Args:
            request: HTTP request
            call_next: Next middleware/handler
            
        Returns:
            HTTP response
        """
        # Skip authentication for public endpoints
        public_paths = [
            "/",
            "/health",
            "/ready", 
            "/metrics",
            "/docs",
            "/redoc",
            "/openapi.json"
        ]
        
        if request.url.path in public_paths:
            return await call_next(request)
        
        # TODO: Implement actual authentication logic
        # For now, just pass through all requests
        
        # Extract user info from headers (if available)
        user_id = request.headers.get("X-User-ID")
        if user_id:
            request.state.user_id = user_id
        
        # Process request
        response = await call_next(request)
        
        return response
