backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# ADK (Agent Development Kit)
# Note: Replace with actual ADK package when available
# adk==1.0.0

# Google Cloud and AI
google-cloud-firestore==2.13.1
google-cloud-storage==2.10.0
google-cloud-aiplatform==1.38.1
google-generativeai==0.3.2
vertexai==1.38.1

# HTTP client for service communication
httpx==0.25.2
aiohttp==3.9.1

# Authentication and security
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Database and caching
redis==5.0.1
asyncpg==0.29.0  # For future PostgreSQL support

# Monitoring and logging
prometheus-client==0.19.0
structlog==23.2.0
python-json-logger==2.0.7

# Image processing
Pillow==10.1.0
opencv-python-headless==4.8.1.78

# Data processing
pandas==2.1.4
numpy==1.25.2

# Environment and configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2  # For testing HTTP clients
pytest-mock==3.12.0

# Development tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Async utilities
asyncio-mqtt==0.16.1  # For future MQTT support
tenacity==8.2.3  # Retry logic

# File handling
aiofiles==23.2.1

# Validation and serialization
marshmallow==3.20.1
jsonschema==4.20.0

# Rate limiting
slowapi==0.1.9

# CORS and middleware
python-cors==1.7.0

# Temporary ADK simulation (remove when real ADK is available)
openai==1.3.7  # For agent simulation
anthropic==0.7.8  # For agent simulation
